:recycle: refactor: Modularizando o código O códio foi modularizado afim de ser mais fácil de fazer manutenção e de compreender
