--- utils.h
@@ -1,12 +1,0 @@
-#ifndef UTILS_H
-#define UTILS_H
-
-// Limitador de quantidade de vértices do grafo
-#define MAX_VERTICES 100
-// Ausência de arestas (distância infinita)
-#define INFINITO 999999 
-
-// Estrutura padrão de uma aresta (ponderada)
-int prim(int grafo[MAX_VERTICES][MAX_VERTICES], int num_vertices);
-
-#endif
