+++ utils/utils.h
@@ -1,0 +1,7 @@
+#ifndef UTILS_H
+#define UTILS_H
+
+double calcularMedia(float *custos, int n);
+double calcularDesvioPadrao(float *custos, int n);
+
+#endif
